pyproject.toml
[tool.poetry]
name = "trad"
version = "0.1.0"
description = ""
authors = ["alexe0110"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
pydantic = "^1.10.7"
uvicorn = "^0.21.1"
redis = "^4.5.4"
sqlalchemy = "1.4.49"
alembic = "^1.10.4"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^11.0.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.0"}
tavern = "^2.0.7"
werkzeug = "^2.3.3"
schemathesis = "^3.19.1"
allure-pytest = "^2.13.2"
black = "^23.12.0"
isort = "^5.13.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
line-length = 120
